+++ src/config.js
@@ -1,3 +1,5 @@
-export const config = { // eslint-disable-line import/prefer-default-export
+const config = {
   panoptesAppId: '24ad5676d5d25c6aa850dc5d5f63ec8c03dbc7ae113b6442b8571fce6c5b974c', // test-rog project for dev
 };
+
+export default config;
